requirements.txt
disnake
deta
python-dotenv
markdownify==0.6.1